# dedups packets using an atomic bloom filter