#### Problem
Current Deduper implementation uses many bits per entry:
https://github.com/solana-labs/solana/blob/65cd55261/perf/src/deduper.rs#L70-L73
and may be saturated quickly. It also lacks api to specify desired false positive rate.



#### Summary of Changes
The commit instead uses an atomic bloom filter with K hash functions. The false positive rate is obtained by tracking popcount of bits.


Note that with K=2, this is 2x slower than the old deduper, because the old deduper had only one hasher. On the other hand because we only consume 2 bits per entry, this gets saturated less quickly and would need less frequent resets.