@t-nelson 2023-03-16
could this not have been implemented as a new type rather than a complete replacement of `Deduper`?  as a b- math student, it's difficult to reason about the statistical differences between the two and no benchmarking results have been provided

were any existing atomic bloom filter implementations considered before we rolled our own?