@behzadnouri 2023-03-17
> and no benchmarking results have been provided

I already mentioned in the PR description that:
> Note that with K=2, this is 2x slower than the old deduper, because the old deduper had only one hasher. On the other hand because we only consume 2 bits per entry, this gets saturated less quickly and would need less frequent resets.

Also discussing with Anatoly, the throughput is still good enough and not the bottleneck in the pipeline.
Benchmark results:
Before:
```
test bench_dedup_baseline           ... bench:          38 ns/iter (+/- 0)
test bench_dedup_diff_big_packets   ... bench:     443,149 ns/iter (+/- 1,651)
test bench_dedup_diff_small_packets ... bench:     149,047 ns/iter (+/- 591)
test bench_dedup_reset              ... bench:     303,878 ns/iter (+/- 767)
test bench_dedup_same_big_packets   ... bench:     363,841 ns/iter (+/- 657)
test bench_dedup_same_small_packets ... bench:     102,606 ns/iter (+/- 349)
```
with bloom:
```
test bench_dedup_baseline           ... bench:          39 ns/iter (+/- 0)
test bench_dedup_diff_big_packets   ... bench:     859,961 ns/iter (+/- 2,044)
test bench_dedup_diff_small_packets ... bench:     266,766 ns/iter (+/- 1,106)
test bench_dedup_reset              ... bench:     119,598 ns/iter (+/- 603)
test bench_dedup_same_big_packets   ... bench:     720,799 ns/iter (+/- 1,806)
test bench_dedup_same_small_packets ... bench:     189,766 ns/iter (+/- 499)
```